PDP11: -Wall warning reductions
pdp11_dsk.c: Change wc's type from int32 to size_t to eliminate type
  slicing/truncation warning.

pdp11_ddcmp.h, pdp11_dmc.c, pdp11_dup.c: Refactor the DDCMP packet
  construction and transmit functions to where they're used
  (pdp11_dmc.c), and out of the pdp11_ddcmp.h header. Eliminates
  "defined but not used" function warnings.

+++ PDP10/pdp6_dsk.c
@@ -260,7 +260,7 @@ dsk_svc (UNIT *uptr)
    int           ctlr  = (dsk_addr >> 16) & 03;
    int           cyl;
    int           sec;
-   int           wc;
+   size_t        wc;
    uint64        data;
    DEVICE       *dptr;
